fix: enable tmux passthru of images if present
I believe this will allow the terminal to read if its passing through tmux...if tmux is configured to emit user env vars which i believe is a user config option. I cant test this on my own but passing 'enable' broke sending plots for me on windows, but leaving in 'detect' with the flag does not.